Likelihood-based statistics of fit for unitquantreg objects.

Computes the likelihood-based statistics (Neg2LogLike, AIC, BIC and HQIC) from unitquantreg objects.

likelihood_stats(..., lt = NULL)

# S3 method for likelihood_stats
print(x, ...)

Arguments

...
unitquantreg objects separated by commas. Not use in print method.

lt
a list with one or more unitquantreg objects.

x
object of class likelihood_stats obtained from likelihood_stats function.

Value

A list with class "likelihood_stats" containing the following components:

call

the matched call.

stats

ordered matrix according AIC value containg the likelihood based statistics.

Details

Neg2LogLike: The log-likelihood is reported as $$Neg2LogLike= -2\log(L)$$

AIC: The Akaike's information criterion (AIC) is defined as $$AIC = -2\log(L)+2p$$

BIC: The Schwarz Bayesian information criterion (BIC) is defined as $$BIC = -2\log(L) + p\log(n)$$

HQIC: The Hannan and Quinn information criterion (HQIC) is defined as $$HQIC = -2\log(L) + 2p\log[\log(n)]$$ where \(L\) is the likelihood function.

References

Akaike, H. (1974). A new look at the statistical model identification. IEEE Transaction on Automatic Control, 19(6), 716--723.

Hannan, E. J. and Quinn, B. G. (1979). The determination of the order of an autoregression. Journal of the Royal Statistical Society, Series B, 41(2), 190--195.

Schwarz, G. (1978). Estimating the dimension of a model. Annals of Statistics, 6(2), 461--464.

Author

André F. B. Menezes andrefelipemaringa@gmail.com

Josmar Mazucheli jmazucheli@gmail.com

Examples

data(water, package = "unitquantreg")

models  <- c("johnsonsb", "kum", "leeg", "ubs", "uburrxii", "uchen", "ughne", "ughnx",
"ugompertz", "ulogistic", "uweibull")
fits <- lapply(models, function(M) unitquantreg(formula = phpws ~ mhdi + incpc +
region + log(pop),tau = 0.5, data = water, family = M))

ans <- likelihood_stats(lt = fits)
ans

#> 
#>  Likelihood-based statistics of fit for unit quantile regression models 
#> 
#> Call:  likelihood_stats(lt = fits)
#> 
#>                                    Neg2LogLike AIC       BIC       HQIC     
#> unit-Logistic                      -8443.029   -8431.029 -8394.140 -8417.855
#> Johnson-SB                         -8238.405   -8226.405 -8189.516 -8213.232
#> unit-Burr-XII                      -8096.114   -8084.114 -8047.225 -8070.941
#> unit-Weibull                       -7994.077   -7982.077 -7945.188 -7968.903
#> unit-Birnbaum-Saunders             -7955.990   -7943.990 -7907.101 -7930.817
#> unit-Gompertz                      -7809.055   -7797.055 -7760.166 -7783.882
#> log-extended Exponential-Geometric -7677.034   -7665.034 -7628.145 -7651.860
#> unit-Generalized Half-Normal-E     -7496.966   -7484.966 -7448.077 -7471.792
#> Kumaraswamy                        -7188.498   -7176.498 -7139.609 -7163.324
#> unit-Chen                          -7016.542   -7004.542 -6967.653 -6991.368
#> unit-Generalized Half-Normal-X     -6965.957   -6953.957 -6917.068 -6940.783