Prediction method for unitquantreg class

Extract various types of predictions from unit quantile regression models.

# S3 method for unitquantreg
predict(
  object,
  newdata,
  type = c("link", "quantile", "shape", "terms"),
  interval = c("none", "confidence"),
  level = 0.95,
  se.fit = FALSE,
  ...
)

Arguments

object

fitted model object of class unitquantreg.

newdata

optionally, a data frame in which to look for variables with which to predict. If omitted, the original observations are used.

type

character indicating type of predictions. The options are link, quantile, shape and terms.

interval

type of interval desired. The options are none and confidence. The "terms" option returns a matrix giving the fitted values of each term in the model formula on the linear predictor scale.

level

coverage probability for the confidence intervals. Default is 0.95.

se.fit

logical. If TRUE return the asymptotic standard errors.

...

currently not used.

Author

André F. B. Menezes