Overview of the unitquantreg package

The unitquantreg R package implements the probability density function, quantile function, cumulative distribution function and random number generation function for 11 distributions with support on the unit-interval. Also, unitquantreg also implements a general infrastructure for parametric quantile regression model based on these distributions.

Details

johnsonsb: implements the [d-q-p-r]johnsonsb function for the Johnson SB distribution.

kum: implements the [d-q-p-r]kum function for the Kumaraswamy distribution.

leeg: implements the [d-q-p-r]leeg function for the Log-extended exponential-geometric distribution.

ubs: implements the [d-q-p-r]ubs function for the unit-Birnbaum-Saunders distribution.

uburrxii: implements the [d-q-p-r]uburrxii function for the unit-Burr-XII distribution.

uchen: implements the [d-q-p-r]uchen function for the unit-Chen distribution.

ughne: implements the [d-q-p-r]ughne function for the unit-Half-Normal-E distribution.

ughnx: implements the [d-q-p-r]ughnx function for the unit-Half-Normal-X distribution.

ugompertz: implements the [d-q-p-r]ugompertz function for the unit-Gompertz distribution.

ulogistic: implements the [d-q-p-r]ulogistic function for the unit-Logistic distribution.

uweibull: implements the [d-q-p-r]uweibull function for the unit-Weibull distribution.

Author

André F. B. Menezes andrefelipemaringa@gmail.com

Josmar Mazucheli jmazucheli@gmail.com