Control parameters for unit quantile regression

Auxiliary function that control fitting of unit quantile regression models using unitquantreg.

unitquantreg.control(
  method = "BFGS",
  hessian = FALSE,
  gradient = TRUE,
  maxit = 5000,
  factr = 1e+07,
  reltol = sqrt(.Machine$double.eps),
  trace = 0L,
  starttests = FALSE,
  dowarn = FALSE,
  ...
)

Arguments

method
string. Specify the method argument passed to optimx.

hessian
logical. Should use the numerically Hessian matrix to compute variance-covariance? Default is FALSE, i.e., use the analytic Hessian.

gradient
logical. Should use the analytic gradient? Default is TRUE.

maxit
integer. Specify the maximal number of iterations passed to optimx.

factr
numeric.Controls the convergence of the "L-BFGS-B" method.

reltol
numeric. Relative convergence tolerance passed to optimx.

trace
non-negative integer. If positive, tracing information on the progress of the optimization is produced.

starttests
logical. Should optimx run tests of the functions and parameters? Default is FALSE.

dowarn
logical. Show warnings generated by optimx? Default is FALSE.

...
arguments passed to optimx.

Value

A list with components named as the arguments.

Details

The control argument of unitquantreg uses the arguments of unitquantreg.control. In particular, the parameters in unitquantreg are estimated by maximum likelihood using the optimx, which is a general-purpose optimization wrapper function that calls other R tools for optimization, including the existing optim function. The main advantage of optimx is to unify the tools allowing a number of different optimization methods and provide sanity checks.

References

Nash, J. C. and Varadhan, R. (2011). Unifying Optimization Algorithms to Aid Software System Users: optimx for R., Journal of Statistical Software, 43(9), 1--14.

See also

Author

André F. B. Menezes

Examples

data(water)
# Fitting using the analytical gradient
fit_gradient <- unitquantreg(formula = phpws ~ mhdi + incpc + region + log(pop),
                             data = water, tau = 0.5, family = "uweibull",
                             control = unitquantreg.control(gradient = TRUE, trace = 1))

#> fn is  fn 
#> Looking for method =  BFGS 
#> Method:  BFGS 
#> initial  value -3888.898517 
#> iter  10 value -3997.038111
#> final  value -3997.038287 
#> converged
#> Post processing for method  BFGS 
#> Successful convergence! 
#> Compute Hessian approximation at finish of  BFGS 
#> Compute gradient approximation at finish of  BFGS 
#> Save results from method  BFGS 
#> $par
#>   (Intercept)          mhdi         incpc        region      log(pop) 
#> -6.5144974552 11.8261951558  0.0002600424 -0.2699077622  0.1054679766 
#>   (Intercept) 
#>  1.1882680569 
#> 
#> $value
#> [1] -3997.038
#> 
#> $message
#> NULL
#> 
#> $convcode
#> [1] 0
#> 
#> $fevals
#> function 
#>       69 
#> 
#> $gevals
#> gradient 
#>       11 
#> 
#> $nitns
#> [1] NA
#> 
#> $kkt1
#> [1] TRUE
#> 
#> $kkt2
#> [1] FALSE
#> 
#> $xtimes
#> user.self 
#>      0.04 
#> 
#> Assemble the answers

# Fitting without using the analytical gradient
fit_nogradient <- unitquantreg(formula = phpws ~ mhdi + incpc + region + log(pop),
                               data = water, tau = 0.5, family = "uweibull",
                               control = unitquantreg.control(gradient = FALSE, trace = 1))

#> fn is  fn 
#> Looking for method =  BFGS 
#> Method:  BFGS 
#> initial  value -3888.898517 
#> final  value -3966.629595 
#> converged
#> Post processing for method  BFGS 
#> Successful convergence! 
#> Compute Hessian approximation at finish of  BFGS 
#> Compute gradient approximation at finish of  BFGS 
#> Save results from method  BFGS 
#> $par
#>   (Intercept)          mhdi         incpc        region      log(pop) 
#> -6.3401686653 11.5056825768 -0.0001163519 -0.3108585471  0.1277500383 
#>   (Intercept) 
#>  1.2561924517 
#> 
#> $value
#> [1] -3966.63
#> 
#> $message
#> NULL
#> 
#> $convcode
#> [1] 0
#> 
#> $fevals
#> function 
#>      175 
#> 
#> $gevals
#> gradient 
#>        9 
#> 
#> $nitns
#> [1] NA
#> 
#> $kkt1
#> [1] FALSE
#> 
#> $kkt2
#> [1] FALSE
#> 
#> $xtimes
#> user.self 
#>     0.138 
#> 
#> Assemble the answers

# Compare estimated coefficients
cbind(gradient = coef(fit_gradient), no_gradient = coef(fit_nogradient))

#>                  gradient   no_gradient
#> (Intercept) -6.5144974552 -6.3401686653
#> mhdi        11.8261951558 11.5056825768
#> incpc        0.0002600424 -0.0001163519
#> region      -0.2699077622 -0.3108585471
#> log(pop)     0.1054679766  0.1277500383
#> theta        1.1882680569  1.2561924517