Vuong test

Performs Vuong test between two fitted objects of class unitquantreg

vuong.test(object1, object2, alternative = c("two.sided", "less", "greater"))

Arguments

object1, object2
objects of class unitquantreg containing the fitted models.

alternative
indicates the alternative hypothesis and must be one of "two.sided" (default), "less", or "greater". You can specify just the initial letter of the value, but the argument name must be given in full. See ‘Details’ for the meanings of the possible values.

Value

A list with class "htest" containing the following components:

statistic

the value of the test statistic.

p.value

the p-value of the test.

alternative

a character string describing the alternative hypothesis.

method

a character string with the method used.

data.name

a character string ginven the name of families models under comparison.

Details

The statistic of Vuong likelihood ratio test for compare two non-nested regression models is defined by $$T = \frac{1}{\widehat{\omega}^2\,\sqrt{n}}\,\sum_{i=1}^{n}\, \log\frac{f(y_i \mid \bm{x}_i, \widehat{\bm{\theta}})}{ g(y_i \mid \bm{x}_i,\widehat{\bm{\gamma}})}$$ where $$\widehat{\omega}^2 = \frac{1}{n}\,\sum_{i=1}^{n}\,\left(\log \frac{f(y_i \mid \bm{x}_i, \widehat{\bm{\theta}})}{g(y_i \mid \bm{x}_i, \widehat{\bm{\gamma}})}\right)^2 - \left[\frac{1}{n}\,\sum_{i=1}^{n}\,\left(\log \frac{f(y_i \mid \bm{x}_i, \widehat{\bm{\theta}})}{ g(y_i \mid \bm{x}_i, \widehat{\bm{\gamma}})}\right)\right]^2$$ is an estimator for the variance of \(\frac{1}{\sqrt{n}}\,\displaystyle\sum_{i=1}^{n}\,\log\frac{f(y_i \mid \bm{x}_i, \widehat{\bm{\theta}})}{g(y_i \mid \bm{x}_i, \widehat{\bm{\gamma}})}\), \(f(y_i \mid \bm{x}_i, \widehat{\bm{\theta}})\) and \(g(y_i \mid \bm{x}_i, \widehat{\bm{\gamma}})\) are the corresponding rival densities evaluated at the maximum likelihood estimates.

When \(n \rightarrow \infty\) we have that \(T \rightarrow N(0, 1)\) in distribution. Therefore, at \(\alpha\%\) level of significance the null hypothesis of the equivalence of the competing models is rejected if \(|T| > z_{\alpha/2}\), where \(z_{\alpha/2}\) is the \(\alpha/2\) quantile of standard normal distribution.

In practical terms, \(f(y_i \mid \bm{x}_i, \widehat{\bm{\theta}})\) is better (worse) than \(g(y_i \mid \bm{x}_i, \widehat{\bm{\gamma}})\) if \(T>z_{\alpha/2}\) (or \(T< -z_{\alpha/2}\)).

References

Vuong, Q. (1989). Likelihood ratio tests for model selection and non-nested hypotheses. Econometrica, 57(2), 307--333.

Author

André F. B. Menezes andrefelipemaringa@gmail.com

Josmar Mazucheli jmazucheli@gmail.com

Examples

data(water, package = "unitquantreg")

fit_uweibull <- unitquantreg(formula = phpws ~ mhdi + incpc + region + log(pop),
tau = 0.5, data = water, family = "uweibull")
fit_kum <- unitquantreg(formula = phpws ~ mhdi + incpc + region + log(pop),
tau = 0.5, data = water, family = "kum")

ans <- vuong.test(object1 = fit_uweibull, object2 = fit_kum)
ans

#> 
#> 	Vuong likelihood ratio test for non-nested models (unit-Weibull versus
#> 	Kumaraswamy)
#> 
#> data:  unit-Weibull versus Kumaraswamy
#> T_LR = 7.2253, p-value = 4.999e-13
#> 

str(ans)

#> List of 4
#>  $ statistic: Named num 7.23
#>   ..- attr(*, "names")= chr "T_LR"
#>  $ p.value  : Named num 5e-13
#>   ..- attr(*, "names")= chr "T_LR"
#>  $ method   : chr "Vuong likelihood ratio test for non-nested models (unit-Weibull versus Kumaraswamy)"
#>  $ data.name: chr "unit-Weibull versus Kumaraswamy"
#>  - attr(*, "class")= chr "htest"