Access to piped water supply data set

the access of people in households with piped water supply in the cities of Brazil from the Southeast and Northeast regions. Information obtained during the census of 2010.

data(water, package = "unitquantreg")

Format

data.frame with 3457 observations and 5 columns:

phpws: the proportion of households with piped water supply.

mhdi: municipal human development index.

incpc: per capita income.

region: 0 for Southeast, 1 for Northeast.

pop: total population.

References

Mazucheli, J., Menezes, A. F. B., Fernandes, L. B., Oliveira, R. P. and Ghitany, M. E., (2020). The unit-Weibull distribution as an alternative to the Kumaraswamy distribution for the modeling of quantiles conditional on covariates. Jounal of Applied Statistics, 47(6), 954--974.

Author

André Felipe Menezes andrefelipemaringa@gmail.com