Percentage of body fat data set

The body fat percentage of individuals assisted in a public hospital in Curitiba, Paraná, Brasil.

data(bodyfat, package = "unitquantreg")

Format

A data.frame with 298 observations and 9 columns:

arms: Arms fat percentage.

legs: Legs fat percentage.

body: Body fat percentage.

android: Android fat percentage.

gynecoid: Ginecoid fat percentage.

bmi: Body mass index - 24.71577.

age: Age - 46.00.

sex: Sex of individual. Female or male.

ipaq: Factor variable indicating the sedentary, insufficiently active or active.

Source

http://www.leg.ufpr.br/doku.php/publications:papercompanions:multquasibeta

References

Petterle, R. R., Bonat, W. H., Scarpin, C. T., Jonasson, T., and Borba, V. Z. C., (2020). Multivariate quasi-beta regression models for continuous bounded data. The International Journal of Biostatistics, 1--15, (preprint).

Mazucheli, J., Leiva, V., Alves, B., and Menezes A. F. B., (2021). A new quantile regression for modeling bounded data under a unit Birnbaum-Saunders distribution with applications in medicine and politics. Symmetry, 13(4) 1--21.

Author

André Felipe Menezes andrefelipemaringa@gmail.com

Josmar Mazucheli jmazucheli@gmail.com

Bruna Alves pg402900@uem.br