Methods for unitquantreg and unitquantregs objects

Methods for extracting information from fitted unit quantile regression objects of class unitquantreg and unitquantregs.

# S3 method for unitquantreg
print(x, digits = max(4, getOption("digits") - 3), ...)

# S3 method for unitquantreg
summary(object, correlation = FALSE, ...)

# S3 method for unitquantreg
coef(object, type = c("full", "quantile", "shape"), ...)

# S3 method for unitquantreg
vcov(object, ...)

# S3 method for unitquantreg
logLik(object, ...)

# S3 method for unitquantreg
confint(object, parm, level = 0.95, ...)

# S3 method for unitquantreg
fitted(object, type = c("quantile", "shape", "full"), ...)

# S3 method for unitquantreg
terms(x, type = c("quantile", "shape"), ...)

# S3 method for unitquantreg
model.frame(formula, ...)

# S3 method for unitquantreg
model.matrix(object, type = c("quantile", "shape"), ...)

# S3 method for unitquantreg
update(object, formula., ..., evaluate = TRUE)

# S3 method for unitquantregs
print(x, digits = max(3, getOption("digits") - 3), ...)

# S3 method for unitquantregs
summary(object, digits = max(3, getOption("digits") - 3), ...)

Arguments

digits

minimal number of significant digits.

...

additional argument(s) for methods. Currently not used.

object, x

fitted model object of class unitquantreg.

correlation

logical; if TRUE, the correlation matrix of the estimated parameters is returned and printed. Default is FALSE.

type

character indicating type of fitted values to return.

parm

a specification of which parameters are to be given confidence intervals, either a vector of numbers or a vector of names. If missing, all parameters are considered.

level

the confidence level required.

formula

an R formula.

formula.

Changes to the formula see update.formula for details.

evaluate

If true evaluate the new call else return the call.

Author

André F. B. Menezes