Methods for unitquantreg and unitquantregs objects

Methods for extracting information from fitted regression models objects of class unitquantreg and unitquantregs.

Usage

# S3 method for unitquantreg
print(x, digits = max(4, getOption("digits") - 3), ...)

# S3 method for unitquantreg
summary(object, correlation = FALSE, ...)

# S3 method for unitquantreg
coef(object, type = c("full", "quantile", "shape"), ...)

# S3 method for unitquantreg
vcov(object, ...)

# S3 method for unitquantreg
logLik(object, ...)

# S3 method for unitquantreg
confint(object, parm, level = 0.95, ...)

# S3 method for unitquantreg
fitted(object, type = c("quantile", "shape", "full"), ...)

# S3 method for unitquantreg
terms(x, type = c("quantile", "shape"), ...)

# S3 method for unitquantreg
model.frame(formula, ...)

# S3 method for unitquantreg
model.matrix(object, type = c("quantile", "shape"), ...)

# S3 method for unitquantreg
update(object, formula., ..., evaluate = TRUE)

# S3 method for unitquantregs
print(x, digits = max(3, getOption("digits") - 3), ...)

# S3 method for unitquantregs
summary(object, digits = max(3, getOption("digits") - 3), ...)

Arguments

digits

minimal number of significant digits.

...

additional argument(s) for methods. Currently not used.

object, x

fitted model object of class unitquantreg.

correlation

logical; if TRUE, the correlation matrix of the estimated parameters is returned and printed. Default is FALSE.

type

character indicating type of fitted values to return.

parm

a specification of which parameters are to be given confidence intervals, either a vector of numbers or a vector of names. If missing, all parameters are considered.

level

the confidence level required.

formula

an R formula.

formula.

Changes to the formula see update.formula for details.

evaluate

If true evaluate the new call else return the call.

Value

The summary method gives Wald tests for the regressions coefficients based on the observed Fisher information matrix. As usual the summary

method returns a list with relevant model statistics and estimates, which can be printed using the print method.

The coef, vcov, confint and fitted methods can be use to extract, respectively, the estimated coefficients, the estimated covariance matrix, the Wald confidence intervals, and fitted values.

A logLik method is also provide, then the AIC

function can be use to calculated the Akaike Information Criterion.

The generic methods terms, model.frame, model.matrix, update and are also provided.

Author

André F. B. Menezes

Examples

set.seed(6669)
n <- 200
betas <- c(1, 2)
gamas <- c(-1, 1)
X <- cbind(1, x1 = runif(n))
Z <- cbind(1, z1 = rexp(n))
eta <- drop(X %*% betas)
zeta <- drop(Z %*% gamas)
mu <- exp(eta) / (1 + exp(eta))
theta <- exp(zeta)
data_simulated <- data.frame(x1 = X[, 2], z1 = Z[, 2])
data_simulated$y <- ruweibull(n, mu, theta)

fit_1 <- unitquantreg(formula = y ~ x1 + z1 + I(x1^2) | z1 + x1,
                      data = data_simulated, family = "uweibull",
                      tau = 0.5, link.theta = "log")
fit_1
#> 
#> unit-Weibull quantile regression model 
#> 
#> Call:  unitquantreg(formula = y ~ x1 + z1 + I(x1^2) | z1 + x1, data = data_simulated, 
#>     tau = 0.5, family = "uweibull", link.theta = "log")
#> 
#> Mu coefficients (quantile model with logit link and tau = 0.5): 
#> (Intercept)           x1           z1      I(x1^2)  
#>      1.0628       1.8619      -0.0074       0.1766  
#> 
#> Theta coefficients (shape model with log link):
#> (Intercept)           z1           x1  
#>     -1.0880       1.1546       0.0310  
#> 
summary(fit_1)
#> 
#>  Wald-tests for unit-Weibull quantile regression model
#> 
#> Call:  unitquantreg(formula = y ~ x1 + z1 + I(x1^2) | z1 + x1, data = data_simulated, 
#>     tau = 0.5, family = "uweibull", link.theta = "log")
#> 
#> Mu coefficients: (quantile model with logit link and tau = 0.5): 
#>              Estimate Std. Error Z value Pr(>|z|)    
#> (Intercept)  1.062804   0.094987  11.189  < 2e-16 ***
#> x1           1.861929   0.283241   6.574 4.91e-11 ***
#> z1          -0.007392   0.028406  -0.260    0.795    
#> I(x1^2)      0.176555   0.297773   0.593    0.553    
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Theta coefficients (shape model with log link):
#>             Estimate Std. Error Z value Pr(>|z|)    
#> (Intercept) -1.08800    0.11055  -9.842   <2e-16 ***
#> z1           1.15457    0.06315  18.284   <2e-16 ***
#> x1           0.03102    0.18214   0.170    0.865    
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Residual degrees of freedom: 193
#> Log-likelihood: 248.7657
#> Number of iterations: 95
vcov(fit_1)
#>                       (Intercept)           x1            z1       I(x1^2)
#> (Intercept)          0.0090225395 -0.001350236 -2.166068e-03 -0.0004797596
#> x1                  -0.0013502361  0.080225545 -4.079395e-03 -0.0819164201
#> z1                  -0.0021660676 -0.004079395  8.069201e-04  0.0042674372
#> I(x1^2)             -0.0004797596 -0.081916420  4.267437e-03  0.0886684631
#> (theta)_(Intercept) -0.0009161317  0.001992448  5.906882e-05 -0.0005448296
#> (theta)_z1          -0.0014292331  0.002120325  1.960683e-04 -0.0024169753
#> (theta)_x1           0.0004251515 -0.010413859  6.844629e-04  0.0078392745
#>                     (theta)_(Intercept)    (theta)_z1    (theta)_x1
#> (Intercept)               -9.161317e-04 -0.0014292331  0.0004251515
#> x1                         1.992448e-03  0.0021203253 -0.0104138590
#> z1                         5.906882e-05  0.0001960683  0.0006844629
#> I(x1^2)                   -5.448296e-04 -0.0024169753  0.0078392745
#> (theta)_(Intercept)        1.222181e-02 -0.0026827060 -0.0155467990
#> (theta)_z1                -2.682706e-03  0.0039876716 -0.0005636706
#> (theta)_x1                -1.554680e-02 -0.0005636706  0.0331756164
coef(fit_1)
#>         (Intercept)                  x1                  z1             I(x1^2) 
#>         1.062804147         1.861929411        -0.007392029         0.176555038 
#> (theta)_(Intercept)          (theta)_z1          (theta)_x1 
#>        -1.088002893         1.154571188         0.031024543 
confint(fit_1)
#>                           2.5 %      97.5 %
#> (Intercept)          0.87663295  1.24897534
#> x1                   1.30678697  2.41707185
#> z1                  -0.06306743  0.04828337
#> I(x1^2)             -0.40706834  0.76017842
#> (theta)_(Intercept) -1.30468145 -0.87132434
#> (theta)_z1           1.03080336  1.27833902
#> (theta)_x1          -0.32596672  0.38801581
terms(fit_1)
#> y ~ x1 + z1 + I(x1^2)
#> attr(,"variables")
#> list(y, x1, z1, I(x1^2))
#> attr(,"factors")
#>         x1 z1 I(x1^2)
#> y        0  0       0
#> x1       1  0       0
#> z1       0  1       0
#> I(x1^2)  0  0       1
#> attr(,"term.labels")
#> [1] "x1"      "z1"      "I(x1^2)"
#> attr(,"order")
#> [1] 1 1 1
#> attr(,"intercept")
#> [1] 1
#> attr(,"response")
#> [1] 1
#> attr(,".Environment")
#> <environment: 0x55f2ae03af90>
model.frame(fit_1)
#>                y          x1           z1      I(x1^2)
#> 1   9.165575e-01 0.186969023 0.3500113669 0.034957....
#> 2   9.868665e-01 0.982630566 0.9999789921 0.965562....
#> 3   8.587095e-01 0.086458551 0.5914376779 0.007475....
#> 4   7.833273e-01 0.346918932 1.1668354254 0.120352....
#> 5   9.211845e-01 0.809755139 1.4398415010 0.655703....
#> 6   9.996629e-01 0.049689748 0.3349306095 0.002469....
#> 7   9.220722e-01 0.925444143 0.2149006084 0.856446....
#> 8   8.794125e-01 0.175631653 0.7831245972 0.030846....
#> 9   8.165429e-01 0.232663407 3.8062623926 0.054132....
#> 10  7.518575e-01 0.710020214 0.5408673137 0.504128....
#> 11  8.152554e-01 0.397797565 0.2765215696 0.158242....
#> 12  9.432065e-01 0.807708004 3.8663189465 0.652392....
#> 13  8.793844e-01 0.089368207 0.2178584193 0.007986....
#> 14  9.728462e-01 0.894879045 1.0438791783 0.800808....
#> 15  9.286123e-01 0.795279165 3.5070576899 0.632468....
#> 16  9.970048e-01 0.863024191 1.4187335444 0.744810....
#> 17  8.757918e-01 0.593657707 1.4076589997 0.352429....
#> 18  8.261351e-01 0.315539946 0.3055390203 0.099565....
#> 19  9.142646e-01 0.598491356 1.8553587458 0.358191....
#> 20  9.955551e-01 0.386077943 0.6311159292 0.149056....
#> 21  6.198507e-01 0.837590330 0.4132693079 0.701557....
#> 22  9.942644e-01 0.489420303 0.9951687632 0.239532....
#> 23  9.753077e-01 0.632291778 0.0634132358 0.399792....
#> 24  5.947303e-01 0.665761166 0.0880321707 0.443237....
#> 25  9.240715e-01 0.634687169 1.1459556231 0.402827....
#> 26  7.824778e-01 0.563399079 0.1356634025 0.317418....
#> 27  8.297512e-01 0.606998358 0.1987294136 0.368447....
#> 28  4.162128e-01 0.388154830 0.1171826934 0.150664....
#> 29  8.834186e-01 0.396943687 1.8777734972 0.157564....
#> 30  2.669522e-02 0.601570645 0.4147568326 0.361887....
#> 31  9.524226e-01 0.568227435 0.9988990230 0.322882....
#> 32  9.890949e-01 0.089259226 0.2074156916 0.007967....
#> 33  8.356917e-01 0.827696131 0.0941934938 0.685080....
#> 34  8.632635e-01 0.221380249 0.8565179053 0.049009....
#> 35  8.224770e-01 0.725421773 1.3613610658 0.526236....
#> 36  9.914971e-01 0.875806357 1.6325865226 0.767036....
#> 37  1.920976e-07 0.344145434 0.0077417375 0.118436....
#> 38  7.009985e-01 0.282891725 0.0164678995 0.080027....
#> 39  3.299573e-01 0.482100938 0.4453643625 0.232421....
#> 40  9.624565e-01 0.965404626 0.0622943919 0.932006....
#> 41  4.460129e-01 0.414056941 0.2542727115 0.171443....
#> 42  6.100226e-01 0.166686361 0.4445197392 0.027784....
#> 43  2.729479e-01 0.377412348 0.1216850207 0.142440....
#> 44  9.628692e-01 0.499762728 0.8014264218 0.249762....
#> 45  7.310191e-01 0.148453972 0.7911051223 0.022038....
#> 46  9.411323e-01 0.897227659 2.0749556198 0.805017....
#> 47  9.174306e-01 0.512832679 1.3166291611 0.262997....
#> 48  7.250009e-01 0.184561140 1.5808561853 0.034062....
#> 49  9.999713e-01 0.252116554 0.0078104455 0.063562....
#> 50  9.244295e-01 0.771840765 1.2487127687 0.595738....
#> 51  8.261196e-01 0.364287117 1.2894351594 0.132705....
#> 52  9.857506e-01 0.340136304 0.0401039790 0.115692....
#> 53  8.980515e-01 0.573364529 2.2463480810 0.328746....
#> 54  6.714400e-02 0.366707520 0.5151595543 0.134474....
#> 55  7.522198e-01 0.052030000 3.3339551857 0.002707....
#> 56  8.070648e-01 0.673684733 0.0308079706 0.453851....
#> 57  8.427202e-01 0.319163317 3.3752158655 0.101865....
#> 58  6.214668e-01 0.739821653 0.1913709114 0.547336....
#> 59  6.226780e-01 0.010372438 0.9954462803 0.000107....
#> 60  4.937106e-01 0.273666156 0.2358366116 0.074893....
#> 61  9.232032e-01 0.498639343 1.1104305125 0.248641....
#> 62  6.842785e-01 0.618480631 0.3946195892 0.382518....
#> 63  8.127175e-01 0.089794202 0.2878930965 0.008062....
#> 64  9.176191e-01 0.542464380 1.0117399795 0.294267....
#> 65  7.419907e-01 0.922037082 0.0513743423 0.850152....
#> 66  8.890993e-01 0.743759766 1.2114763641 0.553178....
#> 67  9.993219e-01 0.530479460 0.0957592814 0.281408....
#> 68  8.671032e-01 0.333831934 1.2031948268 0.111443....
#> 69  9.868402e-01 0.303596644 0.6564226476 0.092170....
#> 70  9.727585e-01 0.779462438 1.5412648990 0.607561....
#> 71  8.941420e-01 0.566358134 4.0994636470 0.320761....
#> 72  7.844396e-01 0.246304353 0.7161469047 0.060665....
#> 73  9.675071e-01 0.453433542 0.2259061099 0.205601....
#> 74  5.321506e-01 0.501688002 0.1983205327 0.251690....
#> 75  8.533438e-01 0.806232679 1.3408606583 0.650011....
#> 76  9.921830e-01 0.876363733 0.5246827621 0.768013....
#> 77  7.000852e-01 0.021096749 2.5232515539 0.000445....
#> 78  4.608770e-01 0.032272287 0.9969185525 0.001041....
#> 79  8.983037e-01 0.232501199 0.7629263056 0.054056....
#> 80  9.841387e-01 0.897619089 0.3010476613 0.805720....
#> 81  8.131413e-01 0.647076813 0.3602553625 0.418708....
#> 82  8.904781e-01 0.440142326 2.3106266427 0.193725....
#> 83  9.078375e-01 0.002572564 1.8448869958 6.618084....
#> 84  8.151917e-01 0.120493138 0.5810949933 0.014518....
#> 85  7.431572e-01 0.058961357 1.5592654626 0.003476....
#> 86  8.154090e-01 0.066204072 0.1208873661 0.004382....
#> 87  8.542268e-01 0.799498970 1.1534450966 0.639198....
#> 88  9.994907e-01 0.650936886 0.2991686645 0.423718....
#> 89  9.682277e-03 0.408658780 0.2572600524 0.167001....
#> 90  1.884694e-01 0.194753532 0.1172813175 0.037928....
#> 91  7.721097e-01 0.401432662 1.2309953291 0.161148....
#> 92  8.422719e-01 0.729597670 1.1978400265 0.532312....
#> 93  9.504483e-01 0.056852603 0.6714319056 0.003232....
#> 94  2.600013e-02 0.030152699 0.1501870452 0.000909....
#> 95  9.610609e-01 0.974481139 1.6735626420 0.949613....
#> 96  9.999979e-01 0.977186496 0.1608033315 0.954893....
#> 97  9.502691e-01 0.954225518 2.4926472360 0.910546....
#> 98  8.460857e-01 0.444790284 0.8314279868 0.197838....
#> 99  1.635179e-02 0.489060622 0.1278958158 0.239180....
#> 100 7.746925e-01 0.882137200 0.1766487421 0.778166....
#> 101 9.874747e-01 0.226013829 0.4694456453 0.051082....
#> 102 9.687905e-01 0.574785293 0.9117094887 0.330378....
#> 103 9.629134e-01 0.506822351 0.4851895925 0.256868....
#> 104 2.677891e-01 0.782133990 0.0849171728 0.611733....
#> 105 5.714224e-01 0.209163900 0.7626485154 0.043749....
#> 106 6.966774e-01 0.657817937 0.1899004711 0.432724....
#> 107 9.461209e-01 0.873680374 0.6471857196 0.763317....
#> 108 9.221385e-01 0.540100920 1.0337704802 0.291709....
#> 109 8.803388e-01 0.733993773 1.7190423048 0.538746....
#> 110 9.362426e-01 0.517044459 1.6952882222 0.267334....
#> 111 9.773931e-01 0.080337508 0.2609683272 0.006454....
#> 112 9.737936e-01 0.224163391 0.0385641429 0.050249....
#> 113 8.971640e-01 0.703506615 2.1930679659 0.494921....
#> 114 9.234557e-01 0.056740811 0.5337154740 0.003219....
#> 115 9.040481e-01 0.600613802 1.9924603066 0.360736....
#> 116 8.739895e-01 0.472318480 3.9672686362 0.223084....
#> 117 7.304058e-01 0.146150552 2.0772778093 0.021359....
#> 118 9.259023e-01 0.404561143 1.3023346560 0.163669....
#> 119 7.121815e-01 0.142716534 1.0452781779 0.020368....
#> 120 5.090681e-03 0.467144786 0.3585611408 0.218224....
#> 121 6.842630e-01 0.007013295 1.2138628354 4.918631....
#> 122 8.025847e-01 0.184218900 2.2340967589 0.033936....
#> 123 9.963403e-01 0.147541127 0.1154633668 0.021768....
#> 124 9.831349e-01 0.088021788 1.1048275921 0.007747....
#> 125 9.850435e-01 0.026480290 0.2011518313 0.000701....
#> 126 9.851405e-01 0.478321804 0.0967952607 0.228791....
#> 127 5.522159e-01 0.711448257 0.4110677387 0.506158....
#> 128 7.724940e-01 0.065931415 1.0951797673 0.004346....
#> 129 2.587477e-01 0.358760210 0.3134477609 0.128708....
#> 130 5.122473e-01 0.229801849 0.3009640351 0.052808....
#> 131 9.985593e-01 0.829483161 0.0597776654 0.688042....
#> 132 7.726487e-01 0.149649319 3.2078082634 0.022394....
#> 133 9.027368e-01 0.837116161 0.4962523989 0.700763....
#> 134 9.449161e-01 0.939180340 1.8500415347 0.882059....
#> 135 4.023865e-01 0.805764747 0.0901670759 0.649256....
#> 136 9.962187e-01 0.337384161 0.3614371032 0.113828....
#> 137 9.466780e-01 0.039731399 0.4086204194 0.001578....
#> 138 6.361434e-01 0.359805279 0.4463318465 0.129459....
#> 139 9.535443e-01 0.983572983 1.0082496656 0.967415....
#> 140 9.076392e-01 0.605062628 2.1487720895 0.366100....
#> 141 6.816570e-01 0.104715174 1.6723449583 0.010965....
#> 142 9.917160e-01 0.732505961 0.4005157799 0.536564....
#> 143 9.466235e-01 0.936254642 0.8242065929 0.876572....
#> 144 9.997844e-01 0.691116306 0.0004143193 0.477641....
#> 145 9.122788e-01 0.450313489 1.4651377733 0.202782....
#> 146 9.825387e-01 0.292686857 1.7038264330 0.085665....
#> 147 9.604232e-01 0.360300791 0.0947520225 0.129816....
#> 148 9.625227e-01 0.363308167 0.7551526949 0.131992....
#> 149 9.497392e-01 0.317096811 0.2216412218 0.100550....
#> 150 7.665515e-01 0.104577311 1.2840862805 0.010936....
#> 151 9.356557e-01 0.520799419 0.5768095111 0.271232....
#> 152 9.711211e-01 0.527755086 0.7634662343 0.278525....
#> 153 9.392989e-01 0.645029359 2.1764425166 0.416062....
#> 154 9.267575e-01 0.085673880 1.9451826065 0.007340....
#> 155 9.040251e-01 0.047223561 0.1791467804 0.002230....
#> 156 8.008470e-01 0.643734220 1.3201216543 0.414393....
#> 157 9.503112e-01 0.679187522 0.2643965036 0.461295....
#> 158 9.134183e-01 0.272767049 1.4695429271 0.074401....
#> 159 6.253765e-01 0.349933921 0.3515539439 0.122453....
#> 160 8.260694e-01 0.786879016 0.9217302929 0.619178....
#> 161 7.137758e-01 0.863003796 0.4296103134 0.744775....
#> 162 8.356647e-01 0.289355269 0.3612351617 0.083726....
#> 163 9.538831e-01 0.802319461 0.0009196382 0.643716....
#> 164 9.698993e-01 0.312108386 0.2762877835 0.097411....
#> 165 9.596746e-01 0.902949745 0.9675859129 0.815318....
#> 166 9.312950e-01 0.810553250 0.7759865364 0.656996....
#> 167 7.669823e-01 0.084399728 1.9517250246 0.007123....
#> 168 9.244149e-01 0.787558387 1.1571210892 0.620248....
#> 169 6.975731e-01 0.548679591 0.1482102266 0.301049....
#> 170 5.674733e-01 0.810101728 0.8974790573 0.656264....
#> 171 9.383495e-01 0.352727176 0.2580404789 0.124416....
#> 172 9.562955e-01 0.810738022 1.1214341959 0.657296....
#> 173 9.633416e-01 0.983013049 0.3024825356 0.966314....
#> 174 8.801466e-01 0.532938308 3.2436927555 0.284023....
#> 175 9.360910e-01 0.562832056 1.1238354587 0.316779....
#> 176 1.514796e-01 0.552046580 0.0824825098 0.304755....
#> 177 7.718311e-01 0.437802015 0.8782899528 0.191670....
#> 178 9.799343e-01 0.172412393 0.1656928626 0.029726....
#> 179 8.451078e-01 0.937999086 0.3594762603 0.879842....
#> 180 9.709155e-01 0.569658614 0.5005534682 0.324510....
#> 181 9.567781e-01 0.204832753 0.8992177145 0.041956....
#> 182 8.271873e-01 0.255136649 2.3168141897 0.065094....
#> 183 4.558835e-07 0.105518685 0.1033205660 0.011134....
#> 184 9.809124e-01 0.909110151 0.6950008859 0.826481....
#> 185 9.695906e-01 0.960477552 1.6198831107 0.922517....
#> 186 9.666374e-01 0.725019708 1.0811691116 0.525653....
#> 187 9.998612e-01 0.469636964 0.6233927791 0.220558....
#> 188 8.183341e-01 0.925040355 0.6160188485 0.855699....
#> 189 9.420188e-01 0.810079255 0.0899007317 0.656228....
#> 190 2.439323e-01 0.162681916 0.5579369292 0.026465....
#> 191 9.226259e-01 0.712089584 2.7083689998 0.507071....
#> 192 8.484646e-01 0.308153212 2.4908233563 0.094958....
#> 193 5.486469e-01 0.585052453 0.0497045792 0.342286....
#> 194 9.999936e-01 0.382884969 0.0568050639 0.146600....
#> 195 9.118204e-01 0.744595389 2.5105187799 0.554422....
#> 196 8.477486e-01 0.766998704 0.0533980178 0.588287....
#> 197 2.165422e-01 0.648972880 0.3470138982 0.421165....
#> 198 1.032769e-02 0.031702994 0.5198124344 0.001005....
#> 199 9.630203e-01 0.665849688 0.5936044552 0.443355....
#> 200 2.551827e-01 0.090263051 0.6568195513 0.008147....
model.matrix(fit_1)
#>     (Intercept)          x1           z1      I(x1^2)
#> 1             1 0.186969023 0.3500113669 3.495742e-02
#> 2             1 0.982630566 0.9999789921 9.655628e-01
#> 3             1 0.086458551 0.5914376779 7.475081e-03
#> 4             1 0.346918932 1.1668354254 1.203527e-01
#> 5             1 0.809755139 1.4398415010 6.557034e-01
#> 6             1 0.049689748 0.3349306095 2.469071e-03
#> 7             1 0.925444143 0.2149006084 8.564469e-01
#> 8             1 0.175631653 0.7831245972 3.084648e-02
#> 9             1 0.232663407 3.8062623926 5.413226e-02
#> 10            1 0.710020214 0.5408673137 5.041287e-01
#> 11            1 0.397797565 0.2765215696 1.582429e-01
#> 12            1 0.807708004 3.8663189465 6.523922e-01
#> 13            1 0.089368207 0.2178584193 7.986676e-03
#> 14            1 0.894879045 1.0438791783 8.008085e-01
#> 15            1 0.795279165 3.5070576899 6.324690e-01
#> 16            1 0.863024191 1.4187335444 7.448108e-01
#> 17            1 0.593657707 1.4076589997 3.524295e-01
#> 18            1 0.315539946 0.3055390203 9.956546e-02
#> 19            1 0.598491356 1.8553587458 3.581919e-01
#> 20            1 0.386077943 0.6311159292 1.490562e-01
#> 21            1 0.837590330 0.4132693079 7.015576e-01
#> 22            1 0.489420303 0.9951687632 2.395322e-01
#> 23            1 0.632291778 0.0634132358 3.997929e-01
#> 24            1 0.665761166 0.0880321707 4.432379e-01
#> 25            1 0.634687169 1.1459556231 4.028278e-01
#> 26            1 0.563399079 0.1356634025 3.174185e-01
#> 27            1 0.606998358 0.1987294136 3.684470e-01
#> 28            1 0.388154830 0.1171826934 1.506642e-01
#> 29            1 0.396943687 1.8777734972 1.575643e-01
#> 30            1 0.601570645 0.4147568326 3.618872e-01
#> 31            1 0.568227435 0.9988990230 3.228824e-01
#> 32            1 0.089259226 0.2074156916 7.967209e-03
#> 33            1 0.827696131 0.0941934938 6.850809e-01
#> 34            1 0.221380249 0.8565179053 4.900921e-02
#> 35            1 0.725421773 1.3613610658 5.262367e-01
#> 36            1 0.875806357 1.6325865226 7.670368e-01
#> 37            1 0.344145434 0.0077417375 1.184361e-01
#> 38            1 0.282891725 0.0164678995 8.002773e-02
#> 39            1 0.482100938 0.4453643625 2.324213e-01
#> 40            1 0.965404626 0.0622943919 9.320061e-01
#> 41            1 0.414056941 0.2542727115 1.714432e-01
#> 42            1 0.166686361 0.4445197392 2.778434e-02
#> 43            1 0.377412348 0.1216850207 1.424401e-01
#> 44            1 0.499762728 0.8014264218 2.497628e-01
#> 45            1 0.148453972 0.7911051223 2.203858e-02
#> 46            1 0.897227659 2.0749556198 8.050175e-01
#> 47            1 0.512832679 1.3166291611 2.629974e-01
#> 48            1 0.184561140 1.5808561853 3.406281e-02
#> 49            1 0.252116554 0.0078104455 6.356276e-02
#> 50            1 0.771840765 1.2487127687 5.957382e-01
#> 51            1 0.364287117 1.2894351594 1.327051e-01
#> 52            1 0.340136304 0.0401039790 1.156927e-01
#> 53            1 0.573364529 2.2463480810 3.287469e-01
#> 54            1 0.366707520 0.5151595543 1.344744e-01
#> 55            1 0.052030000 3.3339551857 2.707121e-03
#> 56            1 0.673684733 0.0308079706 4.538511e-01
#> 57            1 0.319163317 3.3752158655 1.018652e-01
#> 58            1 0.739821653 0.1913709114 5.473361e-01
#> 59            1 0.010372438 0.9954462803 1.075875e-04
#> 60            1 0.273666156 0.2358366116 7.489317e-02
#> 61            1 0.498639343 1.1104305125 2.486412e-01
#> 62            1 0.618480631 0.3946195892 3.825183e-01
#> 63            1 0.089794202 0.2878930965 8.062999e-03
#> 64            1 0.542464380 1.0117399795 2.942676e-01
#> 65            1 0.922037082 0.0513743423 8.501524e-01
#> 66            1 0.743759766 1.2114763641 5.531786e-01
#> 67            1 0.530479460 0.0957592814 2.814085e-01
#> 68            1 0.333831934 1.2031948268 1.114438e-01
#> 69            1 0.303596644 0.6564226476 9.217092e-02
#> 70            1 0.779462438 1.5412648990 6.075617e-01
#> 71            1 0.566358134 4.0994636470 3.207615e-01
#> 72            1 0.246304353 0.7161469047 6.066583e-02
#> 73            1 0.453433542 0.2259061099 2.056020e-01
#> 74            1 0.501688002 0.1983205327 2.516909e-01
#> 75            1 0.806232679 1.3408606583 6.500111e-01
#> 76            1 0.876363733 0.5246827621 7.680134e-01
#> 77            1 0.021096749 2.5232515539 4.450728e-04
#> 78            1 0.032272287 0.9969185525 1.041501e-03
#> 79            1 0.232501199 0.7629263056 5.405681e-02
#> 80            1 0.897619089 0.3010476613 8.057200e-01
#> 81            1 0.647076813 0.3602553625 4.187084e-01
#> 82            1 0.440142326 2.3106266427 1.937253e-01
#> 83            1 0.002572564 1.8448869958 6.618084e-06
#> 84            1 0.120493138 0.5810949933 1.451860e-02
#> 85            1 0.058961357 1.5592654626 3.476442e-03
#> 86            1 0.066204072 0.1208873661 4.382979e-03
#> 87            1 0.799498970 1.1534450966 6.391986e-01
#> 88            1 0.650936886 0.2991686645 4.237188e-01
#> 89            1 0.408658780 0.2572600524 1.670020e-01
#> 90            1 0.194753532 0.1172813175 3.792894e-02
#> 91            1 0.401432662 1.2309953291 1.611482e-01
#> 92            1 0.729597670 1.1978400265 5.323128e-01
#> 93            1 0.056852603 0.6714319056 3.232218e-03
#> 94            1 0.030152699 0.1501870452 9.091852e-04
#> 95            1 0.974481139 1.6735626420 9.496135e-01
#> 96            1 0.977186496 0.1608033315 9.548934e-01
#> 97            1 0.954225518 2.4926472360 9.105463e-01
#> 98            1 0.444790284 0.8314279868 1.978384e-01
#> 99            1 0.489060622 0.1278958158 2.391803e-01
#> 100           1 0.882137200 0.1766487421 7.781660e-01
#> 101           1 0.226013829 0.4694456453 5.108225e-02
#> 102           1 0.574785293 0.9117094887 3.303781e-01
#> 103           1 0.506822351 0.4851895925 2.568689e-01
#> 104           1 0.782133990 0.0849171728 6.117336e-01
#> 105           1 0.209163900 0.7626485154 4.374954e-02
#> 106           1 0.657817937 0.1899004711 4.327244e-01
#> 107           1 0.873680374 0.6471857196 7.633174e-01
#> 108           1 0.540100920 1.0337704802 2.917090e-01
#> 109           1 0.733993773 1.7190423048 5.387469e-01
#> 110           1 0.517044459 1.6952882222 2.673350e-01
#> 111           1 0.080337508 0.2609683272 6.454115e-03
#> 112           1 0.224163391 0.0385641429 5.024923e-02
#> 113           1 0.703506615 2.1930679659 4.949216e-01
#> 114           1 0.056740811 0.5337154740 3.219520e-03
#> 115           1 0.600613802 1.9924603066 3.607369e-01
#> 116           1 0.472318480 3.9672686362 2.230847e-01
#> 117           1 0.146150552 2.0772778093 2.135998e-02
#> 118           1 0.404561143 1.3023346560 1.636697e-01
#> 119           1 0.142716534 1.0452781779 2.036801e-02
#> 120           1 0.467144786 0.3585611408 2.182243e-01
#> 121           1 0.007013295 1.2138628354 4.918631e-05
#> 122           1 0.184218900 2.2340967589 3.393660e-02
#> 123           1 0.147541127 0.1154633668 2.176838e-02
#> 124           1 0.088021788 1.1048275921 7.747835e-03
#> 125           1 0.026480290 0.2011518313 7.012057e-04
#> 126           1 0.478321804 0.0967952607 2.287917e-01
#> 127           1 0.711448257 0.4110677387 5.061586e-01
#> 128           1 0.065931415 1.0951797673 4.346952e-03
#> 129           1 0.358760210 0.3134477609 1.287089e-01
#> 130           1 0.229801849 0.3009640351 5.280889e-02
#> 131           1 0.829483161 0.0597776654 6.880423e-01
#> 132           1 0.149649319 3.2078082634 2.239492e-02
#> 133           1 0.837116161 0.4962523989 7.007635e-01
#> 134           1 0.939180340 1.8500415347 8.820597e-01
#> 135           1 0.805764747 0.0901670759 6.492568e-01
#> 136           1 0.337384161 0.3614371032 1.138281e-01
#> 137           1 0.039731399 0.4086204194 1.578584e-03
#> 138           1 0.359805279 0.4463318465 1.294598e-01
#> 139           1 0.983572983 1.0082496656 9.674158e-01
#> 140           1 0.605062628 2.1487720895 3.661008e-01
#> 141           1 0.104715174 1.6723449583 1.096527e-02
#> 142           1 0.732505961 0.4005157799 5.365650e-01
#> 143           1 0.936254642 0.8242065929 8.765728e-01
#> 144           1 0.691116306 0.0004143193 4.776417e-01
#> 145           1 0.450313489 1.4651377733 2.027822e-01
#> 146           1 0.292686857 1.7038264330 8.566560e-02
#> 147           1 0.360300791 0.0947520225 1.298167e-01
#> 148           1 0.363308167 0.7551526949 1.319928e-01
#> 149           1 0.317096811 0.2216412218 1.005504e-01
#> 150           1 0.104577311 1.2840862805 1.093641e-02
#> 151           1 0.520799419 0.5768095111 2.712320e-01
#> 152           1 0.527755086 0.7634662343 2.785254e-01
#> 153           1 0.645029359 2.1764425166 4.160629e-01
#> 154           1 0.085673880 1.9451826065 7.340014e-03
#> 155           1 0.047223561 0.1791467804 2.230065e-03
#> 156           1 0.643734220 1.3201216543 4.143937e-01
#> 157           1 0.679187522 0.2643965036 4.612957e-01
#> 158           1 0.272767049 1.4695429271 7.440186e-02
#> 159           1 0.349933921 0.3515539439 1.224537e-01
#> 160           1 0.786879016 0.9217302929 6.191786e-01
#> 161           1 0.863003796 0.4296103134 7.447756e-01
#> 162           1 0.289355269 0.3612351617 8.372647e-02
#> 163           1 0.802319461 0.0009196382 6.437165e-01
#> 164           1 0.312108386 0.2762877835 9.741164e-02
#> 165           1 0.902949745 0.9675859129 8.153182e-01
#> 166           1 0.810553250 0.7759865364 6.569966e-01
#> 167           1 0.084399728 1.9517250246 7.123314e-03
#> 168           1 0.787558387 1.1571210892 6.202482e-01
#> 169           1 0.548679591 0.1482102266 3.010493e-01
#> 170           1 0.810101728 0.8974790573 6.562648e-01
#> 171           1 0.352727176 0.2580404789 1.244165e-01
#> 172           1 0.810738022 1.1214341959 6.572961e-01
#> 173           1 0.983013049 0.3024825356 9.663147e-01
#> 174           1 0.532938308 3.2436927555 2.840232e-01
#> 175           1 0.562832056 1.1238354587 3.167799e-01
#> 176           1 0.552046580 0.0824825098 3.047554e-01
#> 177           1 0.437802015 0.8782899528 1.916706e-01
#> 178           1 0.172412393 0.1656928626 2.972603e-02
#> 179           1 0.937999086 0.3594762603 8.798423e-01
#> 180           1 0.569658614 0.5005534682 3.245109e-01
#> 181           1 0.204832753 0.8992177145 4.195646e-02
#> 182           1 0.255136649 2.3168141897 6.509471e-02
#> 183           1 0.105518685 0.1033205660 1.113419e-02
#> 184           1 0.909110151 0.6950008859 8.264813e-01
#> 185           1 0.960477552 1.6198831107 9.225171e-01
#> 186           1 0.725019708 1.0811691116 5.256536e-01
#> 187           1 0.469636964 0.6233927791 2.205589e-01
#> 188           1 0.925040355 0.6160188485 8.556997e-01
#> 189           1 0.810079255 0.0899007317 6.562284e-01
#> 190           1 0.162681916 0.5579369292 2.646541e-02
#> 191           1 0.712089584 2.7083689998 5.070716e-01
#> 192           1 0.308153212 2.4908233563 9.495840e-02
#> 193           1 0.585052453 0.0497045792 3.422864e-01
#> 194           1 0.382884969 0.0568050639 1.466009e-01
#> 195           1 0.744595389 2.5105187799 5.544223e-01
#> 196           1 0.766998704 0.0533980178 5.882870e-01
#> 197           1 0.648972880 0.3470138982 4.211658e-01
#> 198           1 0.031702994 0.5198124344 1.005080e-03
#> 199           1 0.665849688 0.5936044552 4.433558e-01
#> 200           1 0.090263051 0.6568195513 8.147418e-03
#> attr(,"assign")
#> [1] 0 1 2 3
update(fit_1, . ~ . -x1)
#> 
#> unit-Weibull quantile regression model 
#> 
#> Call:  unitquantreg(formula = y ~ z1 + I(x1^2) | z1 + x1, data = data_simulated, 
#>     tau = 0.5, family = "uweibull", link.theta = "log")
#> 
#> Mu coefficients (quantile model with logit link and tau = 0.5): 
#> (Intercept)           z1      I(x1^2)  
#>      1.2658       0.0371       2.0538  
#> 
#> Theta coefficients (shape model with log link):
#> (Intercept)           z1           x1  
#>     -1.0971       0.9338       0.3176  
#> 
update(fit_1, . ~ . -z1)
#> 
#> unit-Weibull quantile regression model 
#> 
#> Call:  unitquantreg(formula = y ~ x1 + I(x1^2) | z1 + x1, data = data_simulated, 
#>     tau = 0.5, family = "uweibull", link.theta = "log")
#> 
#> Mu coefficients (quantile model with logit link and tau = 0.5): 
#> (Intercept)           x1      I(x1^2)  
#>      1.0430       1.8246       0.2154  
#> 
#> Theta coefficients (shape model with log link):
#> (Intercept)           z1           x1  
#>     -1.0873       1.1560       0.0374  
#> 
update(fit_1, . ~ . -I(x1^2))
#> 
#> unit-Weibull quantile regression model 
#> 
#> Call:  unitquantreg(formula = y ~ x1 + z1 | z1 + x1, data = data_simulated, 
#>     tau = 0.5, family = "uweibull", link.theta = "log")
#> 
#> Mu coefficients (quantile model with logit link and tau = 0.5): 
#> (Intercept)           x1           z1  
#>      1.0638       2.0258      -0.0161  
#> 
#> Theta coefficients (shape model with log link):
#> (Intercept)           z1           x1  
#>     -1.0851       1.1565       0.0145  
#> 
update(fit_1, . ~ . | . -z1)
#> 
#> unit-Weibull quantile regression model 
#> 
#> Call:  unitquantreg(formula = y ~ x1 + z1 + I(x1^2) | x1, data = data_simulated, 
#>     tau = 0.5, family = "uweibull", link.theta = "log")
#> 
#> Mu coefficients (quantile model with logit link and tau = 0.5): 
#> (Intercept)           x1           z1      I(x1^2)  
#>      1.3993      -2.6083       0.5267       4.1177  
#> 
#> Theta coefficients (shape model with log link):
#> (Intercept)           x1  
#>     -0.7104       0.5783  
#> 
update(fit_1, . ~ . | . -x1)
#> 
#> unit-Weibull quantile regression model 
#> 
#> Call:  unitquantreg(formula = y ~ x1 + z1 + I(x1^2) | z1, data = data_simulated, 
#>     tau = 0.5, family = "uweibull", link.theta = "log")
#> 
#> Mu coefficients (quantile model with logit link and tau = 0.5): 
#> (Intercept)           x1           z1      I(x1^2)  
#>      1.0621       1.8721      -0.0080       0.1687  
#> 
#> Theta coefficients (shape model with log link):
#> (Intercept)           z1  
#>     -1.0741       1.1559  
#>