Methods for unitquantreg and unitquantregs objects

Methods for extracting information from fitted regression models objects of class unitquantreg and unitquantregs.

Usage

# S3 method for unitquantreg
print(x, digits = max(4, getOption("digits") - 3), ...)

# S3 method for unitquantreg
summary(object, correlation = FALSE, ...)

# S3 method for unitquantreg
coef(object, type = c("full", "quantile", "shape"), ...)

# S3 method for unitquantreg
vcov(object, ...)

# S3 method for unitquantreg
logLik(object, ...)

# S3 method for unitquantreg
confint(object, parm, level = 0.95, ...)

# S3 method for unitquantreg
fitted(object, type = c("quantile", "shape", "full"), ...)

# S3 method for unitquantreg
terms(x, type = c("quantile", "shape"), ...)

# S3 method for unitquantreg
model.frame(formula, ...)

# S3 method for unitquantreg
model.matrix(object, type = c("quantile", "shape"), ...)

# S3 method for unitquantreg
update(object, formula., ..., evaluate = TRUE)

# S3 method for unitquantregs
print(x, digits = max(3, getOption("digits") - 3), ...)

# S3 method for unitquantregs
summary(object, digits = max(3, getOption("digits") - 3), ...)

Arguments

digits

minimal number of significant digits.

...

additional argument(s) for methods. Currently not used.

object, x

fitted model object of class unitquantreg.

correlation

logical; if TRUE, the correlation matrix of the estimated parameters is returned and printed. Default is FALSE.

type

character indicating type of fitted values to return.

parm

a specification of which parameters are to be given confidence intervals, either a vector of numbers or a vector of names. If missing, all parameters are considered.

level

the confidence level required.

formula

an R formula.

formula.

Changes to the formula see update.formula for details.

evaluate

If true evaluate the new call else return the call.

Value

The summary method gives Wald tests for the regressions coefficients based on the observed Fisher information matrix. As usual the summary

method returns a list with relevant model statistics and estimates, which can be printed using the print method.

The coef, vcov, confint and fitted methods can be use to extract, respectively, the estimated coefficients, the estimated covariance matrix, the Wald confidence intervals, and fitted values.

A logLik method is also provide, then the AIC

function can be use to calculated the Akaike Information Criterion.

The generic methods terms, model.frame, model.matrix, update and are also provided.

Author

André F. B. Menezes

Examples

set.seed(6669)
n <- 200
betas <- c(1, 2)
gamas <- c(-1, 1)
X <- cbind(1, x1 = runif(n))
Z <- cbind(1, z1 = rexp(n))
eta <- drop(X %*% betas)
zeta <- drop(Z %*% gamas)
mu <- exp(eta) / (1 + exp(eta))
theta <- exp(zeta)
data_simulated <- data.frame(x1 = X[, 2], z1 = Z[, 2])
data_simulated$y <- ruweibull(n, mu, theta)

fit_1 <- unitquantreg(formula = y ~ x1 + z1 + I(x1^2) | z1 + x1,
                      data = data_simulated, family = "uweibull",
                      tau = 0.5, link.theta = "log")
fit_1
#> 
#> unit-Weibull quantile regression model 
#> 
#> Call:  unitquantreg(formula = y ~ x1 + z1 + I(x1^2) | z1 + x1, data = data_simulated, 
#>     tau = 0.5, family = "uweibull", link.theta = "log")
#> 
#> Mu coefficients (quantile model with logit link and tau = 0.5): 
#> (Intercept)           x1           z1      I(x1^2)  
#>      1.0628       1.8619      -0.0074       0.1766  
#> 
#> Theta coefficients (shape model with log link):
#> (Intercept)           z1           x1  
#>     -1.0880       1.1546       0.0310  
#> 
summary(fit_1)
#> 
#>  Wald-tests for unit-Weibull quantile regression model
#> 
#> Call:  unitquantreg(formula = y ~ x1 + z1 + I(x1^2) | z1 + x1, data = data_simulated, 
#>     tau = 0.5, family = "uweibull", link.theta = "log")
#> 
#> Mu coefficients: (quantile model with logit link and tau = 0.5): 
#>              Estimate Std. Error Z value Pr(>|z|)    
#> (Intercept)  1.062804   0.094987  11.189  < 2e-16 ***
#> x1           1.861929   0.283241   6.574 4.91e-11 ***
#> z1          -0.007392   0.028406  -0.260    0.795    
#> I(x1^2)      0.176555   0.297773   0.593    0.553    
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Theta coefficients (shape model with log link):
#>             Estimate Std. Error Z value Pr(>|z|)    
#> (Intercept) -1.08800    0.11055  -9.842   <2e-16 ***
#> z1           1.15457    0.06315  18.284   <2e-16 ***
#> x1           0.03102    0.18214   0.170    0.865    
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Residual degrees of freedom: 193
#> Log-likelihood: 248.7657
#> Number of iterations: 95
vcov(fit_1)
#>                       (Intercept)           x1            z1       I(x1^2)
#> (Intercept)          0.0090225395 -0.001350236 -2.166068e-03 -0.0004797596
#> x1                  -0.0013502361  0.080225545 -4.079395e-03 -0.0819164201
#> z1                  -0.0021660676 -0.004079395  8.069201e-04  0.0042674372
#> I(x1^2)             -0.0004797596 -0.081916420  4.267437e-03  0.0886684631
#> (theta)_(Intercept) -0.0009161317  0.001992448  5.906882e-05 -0.0005448296
#> (theta)_z1          -0.0014292331  0.002120325  1.960683e-04 -0.0024169753
#> (theta)_x1           0.0004251515 -0.010413859  6.844629e-04  0.0078392745
#>                     (theta)_(Intercept)    (theta)_z1    (theta)_x1
#> (Intercept)               -9.161317e-04 -0.0014292331  0.0004251515
#> x1                         1.992448e-03  0.0021203253 -0.0104138590
#> z1                         5.906882e-05  0.0001960683  0.0006844629
#> I(x1^2)                   -5.448296e-04 -0.0024169753  0.0078392745
#> (theta)_(Intercept)        1.222181e-02 -0.0026827060 -0.0155467990
#> (theta)_z1                -2.682706e-03  0.0039876716 -0.0005636706
#> (theta)_x1                -1.554680e-02 -0.0005636706  0.0331756164
coef(fit_1)
#>         (Intercept)                  x1                  z1             I(x1^2) 
#>         1.062804147         1.861929411        -0.007392029         0.176555038 
#> (theta)_(Intercept)          (theta)_z1          (theta)_x1 
#>        -1.088002893         1.154571188         0.031024543 
confint(fit_1)
#>                           2.5 %      97.5 %
#> (Intercept)          0.87663295  1.24897534
#> x1                   1.30678697  2.41707185
#> z1                  -0.06306743  0.04828337
#> I(x1^2)             -0.40706834  0.76017842
#> (theta)_(Intercept) -1.30468145 -0.87132434
#> (theta)_z1           1.03080336  1.27833902
#> (theta)_x1          -0.32596672  0.38801581
terms(fit_1)
#> y ~ x1 + z1 + I(x1^2)
#> attr(,"variables")
#> list(y, x1, z1, I(x1^2))
#> attr(,"factors")
#>         x1 z1 I(x1^2)
#> y        0  0       0
#> x1       1  0       0
#> z1       0  1       0
#> I(x1^2)  0  0       1
#> attr(,"term.labels")
#> [1] "x1"      "z1"      "I(x1^2)"
#> attr(,"order")
#> [1] 1 1 1
#> attr(,"intercept")
#> [1] 1
#> attr(,"response")
#> [1] 1
#> attr(,".Environment")
#> <environment: 0x55b8165d5aa8>
model.frame(fit_1)[1:5, ]
#>           y         x1        z1      I(x1^2)
#> 1 0.9165575 0.18696902 0.3500114 0.034957....
#> 2 0.9868665 0.98263057 0.9999790 0.965562....
#> 3 0.8587095 0.08645855 0.5914377 0.007475....
#> 4 0.7833273 0.34691893 1.1668354 0.120352....
#> 5 0.9211845 0.80975514 1.4398415 0.655703....
model.matrix(fit_1)[1:5, ]
#>   (Intercept)         x1        z1     I(x1^2)
#> 1           1 0.18696902 0.3500114 0.034957415
#> 2           1 0.98263057 0.9999790 0.965562828
#> 3           1 0.08645855 0.5914377 0.007475081
#> 4           1 0.34691893 1.1668354 0.120352745
#> 5           1 0.80975514 1.4398415 0.655703385
update(fit_1, . ~ . -x1)
#> 
#> unit-Weibull quantile regression model 
#> 
#> Call:  unitquantreg(formula = y ~ z1 + I(x1^2) | z1 + x1, data = data_simulated, 
#>     tau = 0.5, family = "uweibull", link.theta = "log")
#> 
#> Mu coefficients (quantile model with logit link and tau = 0.5): 
#> (Intercept)           z1      I(x1^2)  
#>      1.2658       0.0371       2.0538  
#> 
#> Theta coefficients (shape model with log link):
#> (Intercept)           z1           x1  
#>     -1.0971       0.9338       0.3176  
#> 
update(fit_1, . ~ . -z1)
#> 
#> unit-Weibull quantile regression model 
#> 
#> Call:  unitquantreg(formula = y ~ x1 + I(x1^2) | z1 + x1, data = data_simulated, 
#>     tau = 0.5, family = "uweibull", link.theta = "log")
#> 
#> Mu coefficients (quantile model with logit link and tau = 0.5): 
#> (Intercept)           x1      I(x1^2)  
#>      1.0430       1.8246       0.2154  
#> 
#> Theta coefficients (shape model with log link):
#> (Intercept)           z1           x1  
#>     -1.0873       1.1560       0.0374  
#> 
update(fit_1, . ~ . -I(x1^2))
#> 
#> unit-Weibull quantile regression model 
#> 
#> Call:  unitquantreg(formula = y ~ x1 + z1 | z1 + x1, data = data_simulated, 
#>     tau = 0.5, family = "uweibull", link.theta = "log")
#> 
#> Mu coefficients (quantile model with logit link and tau = 0.5): 
#> (Intercept)           x1           z1  
#>      1.0638       2.0258      -0.0161  
#> 
#> Theta coefficients (shape model with log link):
#> (Intercept)           z1           x1  
#>     -1.0851       1.1565       0.0145  
#> 
update(fit_1, . ~ . | . -z1)
#> 
#> unit-Weibull quantile regression model 
#> 
#> Call:  unitquantreg(formula = y ~ x1 + z1 + I(x1^2) | x1, data = data_simulated, 
#>     tau = 0.5, family = "uweibull", link.theta = "log")
#> 
#> Mu coefficients (quantile model with logit link and tau = 0.5): 
#> (Intercept)           x1           z1      I(x1^2)  
#>      1.3993      -2.6083       0.5267       4.1177  
#> 
#> Theta coefficients (shape model with log link):
#> (Intercept)           x1  
#>     -0.7104       0.5783  
#> 
update(fit_1, . ~ . | . -x1)
#> 
#> unit-Weibull quantile regression model 
#> 
#> Call:  unitquantreg(formula = y ~ x1 + z1 + I(x1^2) | z1, data = data_simulated, 
#>     tau = 0.5, family = "uweibull", link.theta = "log")
#> 
#> Mu coefficients (quantile model with logit link and tau = 0.5): 
#> (Intercept)           x1           z1      I(x1^2)  
#>      1.0621       1.8721      -0.0080       0.1687  
#> 
#> Theta coefficients (shape model with log link):
#> (Intercept)           z1  
#>     -1.0741       1.1559  
#>