Pairwise vuong test

Calculate pairwise comparisons between fitted models performing vuong test for objects of class unitquantreg.

pairwise.vuong.test(
  ...,
  lt,
  p.adjust.method = p.adjust.methods,
  alternative = c("two.sided", "less", "greater")
)

Arguments

...
unitquantreg objects separated by commas.

lt
a list with one or more unitquantreg objects.

p.adjust.method
a character string specifying the method for multiple testing adjustment; almost always one of p.adjust.methods. Can be abbreviated.

alternative
indicates the alternative hypothesis and must be one of "two.sided" (default), "less", or "greater". Can be abbreviated.

Value

Object of class "pairwise.htest"

See also

Examples

data(water, package = "unitquantreg")

models  <- c("johnsonsb", "kum", "leeg", "ubs", "uburrxii", "uchen", "ughne", "ughnx",
"ugompertz", "ulogistic", "uweibull")
fits <- lapply(models, function(M) unitquantreg(formula = phpws ~ mhdi + incpc +
region + log(pop),tau = 0.5, data = water, family = M))

ans <- pairwise.vuong.test(lt = fits)
ans

#> 
#> 	Pairwise comparisons using Vuong likelihood ratio test for non-nested models 
#> 
#> data:  unitquantreg(formula = phpws ~ mhdi + incpc + region + log(pop), 
#>     data = water, tau = 0.5, family = M) 
#> 
#>                                    Johnson-SB Kumaraswamy
#> Kumaraswamy                        < 2e-16    -          
#> log-extended Exponential-Geometric < 2e-16    < 2e-16    
#> unit-Birnbaum-Saunders             0.00420    6.2e-06    
#> unit-Burr-XII                      0.61768    3.1e-09    
#> unit-Chen                          < 2e-16    0.64910    
#> unit-Generalized Half-Normal-E     < 2e-16    0.07703    
#> unit-Generalized Half-Normal-X     < 2e-16    0.64910    
#> unit-Gompertz                      1.1e-09    5.4e-06    
#> unit-Logistic                      0.00174    < 2e-16    
#> unit-Weibull                       0.00060    1.9e-11    
#>                                    log-extended Exponential-Geometric
#> Kumaraswamy                        -                                 
#> log-extended Exponential-Geometric -                                 
#> unit-Birnbaum-Saunders             0.26769                           
#> unit-Burr-XII                      0.00484                           
#> unit-Chen                          4.1e-10                           
#> unit-Generalized Half-Normal-E     0.38220                           
#> unit-Generalized Half-Normal-X     0.00024                           
#> unit-Gompertz                      0.64910                           
#> unit-Logistic                      2.9e-16                           
#> unit-Weibull                       0.00267                           
#>                                    unit-Birnbaum-Saunders unit-Burr-XII
#> Kumaraswamy                        -                      -            
#> log-extended Exponential-Geometric -                      -            
#> unit-Birnbaum-Saunders             -                      -            
#> unit-Burr-XII                      0.61768                -            
#> unit-Chen                          1.4e-06                1.3e-09      
#> unit-Generalized Half-Normal-E     0.00099                1.7e-14      
#> unit-Generalized Half-Normal-X     < 2e-16                9.0e-15      
#> unit-Gompertz                      0.53027                2.4e-09      
#> unit-Logistic                      4.2e-11                2.0e-06      
#> unit-Weibull                       1.00000                0.26769      
#>                                    unit-Chen unit-Generalized Half-Normal-E
#> Kumaraswamy                        -         -                             
#> log-extended Exponential-Geometric -         -                             
#> unit-Birnbaum-Saunders             -         -                             
#> unit-Burr-XII                      -         -                             
#> unit-Chen                          -         -                             
#> unit-Generalized Half-Normal-E     0.00015   -                             
#> unit-Generalized Half-Normal-X     1.00000   0.02107                       
#> unit-Gompertz                      5.7e-07   0.00015                       
#> unit-Logistic                      < 2e-16   < 2e-16                       
#> unit-Weibull                       7.9e-14   < 2e-16                       
#>                                    unit-Generalized Half-Normal-X unit-Gompertz
#> Kumaraswamy                        -                              -            
#> log-extended Exponential-Geometric -                              -            
#> unit-Birnbaum-Saunders             -                              -            
#> unit-Burr-XII                      -                              -            
#> unit-Chen                          -                              -            
#> unit-Generalized Half-Normal-E     -                              -            
#> unit-Generalized Half-Normal-X     -                              -            
#> unit-Gompertz                      1.2e-09                        -            
#> unit-Logistic                      < 2e-16                        < 2e-16      
#> unit-Weibull                       8.2e-11                        0.00023      
#>                                    unit-Logistic
#> Kumaraswamy                        -            
#> log-extended Exponential-Geometric -            
#> unit-Birnbaum-Saunders             -            
#> unit-Burr-XII                      -            
#> unit-Chen                          -            
#> unit-Generalized Half-Normal-E     -            
#> unit-Generalized Half-Normal-X     -            
#> unit-Gompertz                      -            
#> unit-Logistic                      -            
#> unit-Weibull                       5.0e-10      
#> 
#> P value adjustment method: holm